Export DB and UI Changes
Added the ability to export the DB to a JSON file so it can be shared.
Added minor UI changes to animations and general look.
Added a base class for the data contexts to inherit from.
Added changes to the ExtensionsModule to make it work better.
Added minor changes to other files.